redirecting routs + create buttons on Welcomepage
